Final design changes
Changes have been made to the design. Now the mobile version really looks like a phone and all the actions that occur on the phone are pushed into a frame that resembles a phone. The same changes affected the ATM, but here everything is pushed into a frame that looks like an ATM. The admin panel did not receive any changes, because it is something more abstract. Some buttons were also removed from the header, which were then duplicated in the phone/ATM.